+++ tests/framework/di/ContainerTest.php
@@ -13,6 +13,7 @@ use yii\di\Instance;
 use yiiunit\framework\di\stubs\Bar;
 use yiiunit\framework\di\stubs\Foo;
 use yiiunit\framework\di\stubs\Qux;
+use yiiunit\framework\di\stubs\QuxInterface;
 use yiiunit\TestCase;
 use yii\validators\NumberValidator;
 
@@ -169,6 +170,8 @@ class ContainerTest extends TestCase
         $this->assertEquals($request, Yii::$app->request);
         $this->assertEquals($response, Yii::$app->response);
 
+
+
     }
 
     public function testAssociativeInvoke()
@@ -213,4 +216,16 @@ class ContainerTest extends TestCase
         $this->assertEquals([1, 5], Yii::$container->resolveCallableDependencies($closure, ['a' => 1, 'b' => 5]));
         $this->assertEquals([1, 5], Yii::$container->resolveCallableDependencies($closure, [1, 5]));
     }
+
+    public function testOptionalDependencies()
+    {
+        $container = new Container;
+        // Test optional unresolvable dependency.
+        $closure = function(QuxInterface $test = null) {
+            return $test;
+        };
+        $this->assertNull($container->invoke($closure));
+    }
+
+
 }